requirements.txt
kivy>=2.3.0
kivymd>=1.1.1
pydantic>=2.7.0
cryptography>=42.0.0
keyring>=24.0.0
python-dotenv>=1.0.1
matplotlib>=3.8.0
reportlab>=4.0.8
python-docx>=1.0.0
